Code/AWS/IoTSiteWise/requirements.txt
aws-cdk-lib==2.71.0
constructs>10.0.0
requests